+++ engine-types/src/parameters/silo.rs
@@ -1,15 +1,20 @@
 use crate::account_id::AccountId;
 use crate::borsh::{self, BorshDeserialize, BorshSerialize};
-use crate::types::{Address, Wei};
+use crate::types::{Address, EthGas};
 
 #[derive(Debug, Clone, PartialEq, Eq, BorshSerialize, BorshDeserialize)]
-pub struct FixedGasCostArgs {
-    pub cost: Option<Wei>,
+pub struct FixedGasArgs {
+    pub fixed_gas: Option<EthGas>,
 }
 
 #[derive(Debug, Default, Clone, PartialEq, Eq, BorshSerialize, BorshDeserialize)]
 pub struct SiloParamsArgs {
-    pub fixed_gas_cost: Wei,
+    /// Fixed amount of gas per transaction.
+    pub fixed_gas: EthGas,
+    /// EVM address, which is used for withdrawing ERC-20 base tokens in case
+    /// a recipient of the tokens is not in the silo white list.
+    /// Note: the logic described above works only if the fallback address
+    /// is set by `set_silo_params` function. In other words, in Silo mode.
     pub erc20_fallback_address: Address,
 }