+++ engine-types/src/types/wei.rs
@@ -1,7 +1,7 @@
 use crate::fmt::Formatter;
 use crate::types::balance::error;
-use crate::types::Fee;
-use crate::{format, Add, Display, Sub, SubAssign, ToString, U256};
+use crate::types::{EthGas, Fee};
+use crate::{format, Add, Display, Mul, Sub, SubAssign, ToString, U256};
 #[cfg(not(feature = "borsh-compat"))]
 use borsh::{maybestd::io, BorshDeserialize, BorshSerialize};
 #[cfg(feature = "borsh-compat")]
@@ -185,6 +185,14 @@ impl Sub for Wei {
 
     fn sub(self, rhs: Self) -> Self::Output {
         Self(self.0 - rhs.0)
+    }
+}
+
+impl Mul<EthGas> for Wei {
+    type Output = Self;
+
+    fn mul(self, rhs: EthGas) -> Self::Output {
+        Self(self.0 * rhs.as_u256())
     }
 }